requirements.txt
amqp==5.2.0
annotated-types==0.7.0
anyio==4.6.2.post1
argcomplete==3.6.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asyncpg==0.30.0
attrs==24.2.0
bcrypt==4.1.2
Beaker==1.12.1
beautifulsoup4==4.12.3
billiard==4.2.1
blinker==1.7.0
blivet==3.11.0
blivet-gui==2.6.0
boto3==1.37.16
botocore==1.37.16
Brlapi==0.8.5
cached-property==1.5.2
celery==5.3.6
cffi==1.17.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
contourpy==1.3.1
crypt_r==3.13.1
cryptography==43.0.0
cupshelpers==1.0
cycler==0.12.1
dasbus==1.7
dbus-python==1.3.2
defusedxml==0.7.1
distlib==0.3.8
distro==1.9.0
dnf==4.23.0
dnspython==2.7.0
docker==7.1.0
docker-pycreds==0.4.0
dockerpty==0.4.1
docopt==0.6.2
email_validator==2.2.0
fastapi==0.115.11
fastapi-users==14.0.1
fastapi-users-db-sqlalchemy==7.0.0
fedora-third-party==0.10
file-magic==0.4.0
filelock==3.15.4
Flask==3.0.3
fonttools==4.56.0
fros==1.1
fsspec==2025.3.0
gitdb==4.0.12
GitPython==3.1.44
google-re2==1.1.20240702
greenlet==3.1.1
h11==0.14.0
httpie==3.2.4
humanize==4.9.0
idna==3.7
iso639==0.1.4
itsdangerous==2.1.2
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.0.2
jeepney==0.8.0
Jinja2==3.1.6
jmespath==1.0.1
jsonschema==4.19.1
jsonschema-specifications==2023.11.2
keyring==25.6.0
kiwisolver==1.4.8
kombu==5.3.7
langtable==0.0.68
libcomps==0.1.21
libdnf==0.74.0
louis==3.28.0
lxml==5.2.1
makefun==1.15.6
Mako==1.2.3
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.10.1
mdurl==0.1.2
mercurial==6.8.2
mod_wsgi==5.0.2
more-itertools==10.1.0
mpmath==1.3.0
multidict==6.0.5
networkx==3.4.2
nftables==0.1
noise==1.2.2
numpy==2.1.1
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
olefile==0.47
opencv-python==4.10.0.84
packaging==24.2
pandas==2.2.3
paramiko==3.5.1
Paste==3.10.1
perf==0.1
pexpect==4.9.0
pid==2.2.3
pillow==11.0.0
platformdirs==4.2.2
ply==3.11
productmd==1.43
prompt_toolkit==3.0.41
psutil==7.0.0
psycopg2==2.9.10
ptyprocess==0.7.0
pwdlib==0.2.1
pwquality==1.4.5
py-cpuinfo==9.0.0
pyasn1==0.6.0
pycairo==1.25.1
pycparser==2.20
pycrypto==2.6.1
pycups==2.0.4
pydantic==2.10.6
pydantic_core==2.27.2
pyenchant==3.2.2
Pygments==2.18.0
PyGObject==3.48.2
pyinotify==0.9.6
PyJWT==2.10.1
pykickstart==3.58
PyNaCl==1.5.0
pyOpenSSL==24.2.1
pyparsing==3.1.2
pyparted==3.13.0
PySocks==1.7.1
python-augeas==1.1.0
python-dateutil==2.8.2
python-dotenv==1.0.1
python-linux-procfs==0.7.3
python-meh==0.52
python-multipart==0.0.20
python-pam==2.0.2
pytz==2025.1
pyudev==0.24.3
pyxdg==0.27
PyYAML==6.0.1
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-file==2.0.0
requests-ftp==0.3.1
requests-toolbelt==1.0.0
rich==13.7.1
rpds-py==0.19.1
rpm==4.20.1
s3transfer==0.11.4
scipy==1.15.2
seaborn==0.13.2
SecretStorage==3.3.3
selinux @ file:///builddir/build/BUILD/libselinux-3.7-build/libselinux-3.7/src
sepolicy @ file:///builddir/build/BUILD/policycoreutils-3.7-build/selinux-3.7/python/sepolicy
setools==4.5.1
setuptools==76.1.0
shtab==1.7.1
simpleaudio==1.0.4
simpleline==1.9.0
six==1.16.0
smmap==5.0.2
sniffio==1.3.1
sos==4.8.2
soupsieve==2.6
SQLAlchemy==2.0.39
starlette==0.41.2
sympy==1.13.1
systemd-python==235
Tempita==0.5.2
texttable==1.6.7
thop==0.1.1.post2209072238
torch==2.6.0
torchvision==0.21.0
tqdm==4.67.1
triton==3.2.0
typing_extensions==4.12.2
tzdata==2025.1
ultralytics==8.3.93
ultralytics-thop==2.0.14
urllib3==1.26.20
uvicorn==0.34.1
vine==5.1.0
virtualenv==20.21.1
wcwidth==0.2.13
websocket-client==1.7.0
Werkzeug==3.0.6
xkbregistry==0.3
